kleinzeugs

+++ webserver/socket.h
@@ -3,6 +3,7 @@
 
 
 #include "config.h"
+#include <stdint.h>
 
 
 #ifdef WS_WIN32
@@ -16,7 +17,7 @@ typedef int socket_t;
 
 bool socket_create(socket_t *socket);
 void socket_destroy(socket_t socket);
-bool socket_bind(socket_t socket, unsigned port);
+bool socket_bind(socket_t socket, uint16_t port);
 bool socket_accept(socket_t socket, socket_t *accepted);
 bool socket_receive(socket_t socket, void *data, size_t size, size_t *received);
 bool socket_send(socket_t socket, const void *data, size_t size);
